requirements.txt
Pillow
reportlab
pypdf
inflect
streamlit
